requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
msgspec==0.18.6
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
valo_api==2.0.8
yarl==1.18.0
tortoise-orm==0.22.1